update return value to i64

+++ ExampleHandler.h
@@ -7,7 +7,7 @@ namespace tamvm {
 namespace cpp2 {
 class ExampleHandler: public ExampleServiceSvIf {
 public:
-	int32_t get_number(int32_t n) override {
+	int64_t get_number(int32_t n) override {
 		LOG(INFO) << "server: receive " << n;
 		return n + 10;
 	}
